feat: frame validation GOAWAY errors WIP
New Connection.Processor module to validate frames, acting on them if
valid.

Also: Ping now has :stream_id
