+++ lib/connection.ex
@@ -9,6 +9,8 @@ defmodule Kadabra.Connection do
   use GenStage
   require Logger
 
+  import Kernel, except: [send: 2]
+
   alias Kadabra.{
     Config,
     Connection,
@@ -16,43 +18,21 @@ defmodule Kadabra.Connection do
     Encodable,
     Error,
     Frame,
-    Hpack,
-    Socket,
-    Stream,
-    StreamSupervisor
+    Socket
   }
 
-  alias Kadabra.Connection.FlowControl
+  alias Kadabra.Frame.{Goaway, Ping}
 
-  alias Kadabra.Frame.{
-    Continuation,
-    Data,
-    Goaway,
-    Headers,
-    Ping,
-    PushPromise,
-    RstStream,
-    WindowUpdate
-  }
+  alias Kadabra.Connection.{FlowControl, Processor}
 
   @type t :: %__MODULE__{
           buffer: binary,
           config: term,
-          flow_control: term
+          flow_control: term,
+          queue: pid
         }
 
   @type sock :: {:sslsocket, any, pid | {any, any}}
-
-  @type frame ::
-          Data.t()
-          | Headers.t()
-          | RstStream.t()
-          | Frame.Settings.t()
-          | PushPromise.t()
-          | Ping.t()
-          | Goaway.t()
-          | WindowUpdate.t()
-          | Continuation.t()
 
   def start_link(%Config{supervisor: sup} = config) do
     name = via_tuple(sup)
@@ -65,7 +45,7 @@ defmodule Kadabra.Connection do
 
   def init(%Config{supervisor: sup} = config) do
     state = initial_state(config)
-    send(self(), :start)
+    Kernel.send(self(), :start)
     {:consumer, state, subscribe_to: [ConnectionQueue.via_tuple(sup)]}
   end
 
@@ -123,7 +103,7 @@ defmodule Kadabra.Connection do
     bin = flow.stream_id |> Goaway.new() |> Encodable.to_bin()
     Socket.send(config.socket, bin)
 
-    send(config.client, {:closed, config.supervisor})
+    Kernel.send(config.client, {:closed, config.supervisor})
 
     Task.Supervisor.start_child(Kadabra.Tasks, fn ->
       Kadabra.Supervisor.stop(config.supervisor)
@@ -147,99 +127,10 @@ defmodule Kadabra.Connection do
 
   # recv
 
-  @spec recv(frame, t) :: {:noreply, [], t}
-  def recv(%Frame.RstStream{}, state) do
-    Logger.error("recv unstarted stream rst")
-    {:noreply, [], state}
-  end
-
-  def recv(%Frame.Ping{ack: true}, %{config: config} = state) do
-    send(config.client, {:pong, self()})
-    {:noreply, [], state}
-  end
-
-  def recv(%Frame.Ping{ack: false}, %{client: pid} = state) do
-    send(pid, {:ping, self()})
-    {:noreply, [], state}
-  end
-
-  # nil settings means use default
-  def recv(%Frame.Settings{ack: false, settings: nil}, state) do
-    %{flow_control: flow, config: config} = state
-
-    bin = Frame.Settings.ack() |> Encodable.to_bin()
-    Socket.send(config.socket, bin)
-
-    case flow.settings.max_concurrent_streams do
-      :infinite ->
-        GenStage.ask(state.queue, 2_000_000_000)
-
-      max ->
-        to_ask = max - flow.active_stream_count
-        GenStage.ask(state.queue, to_ask)
-    end
-
-    {:noreply, [], state}
-  end
-
-  def recv(%Frame.Settings{ack: false, settings: settings}, state) do
-    %{flow_control: flow, config: config} = state
-    old_settings = flow.settings
-    flow = Connection.FlowControl.update_settings(flow, settings)
-
-    notify_settings_change(config.ref, old_settings, flow)
-
-    pid = Hpack.via_tuple(config.ref, :encoder)
-    Hpack.update_max_table_size(pid, settings.max_header_list_size)
-
-    bin = Frame.Settings.ack() |> Encodable.to_bin()
-    Socket.send(config.socket, bin)
-
-    to_ask = settings.max_concurrent_streams - flow.active_stream_count
-    GenStage.ask(state.queue, to_ask)
-
-    {:noreply, [], %{state | flow_control: flow}}
-  end
-
-  def recv(%Frame.Settings{ack: true}, state) do
-    send_huge_window_update(state.config.socket)
-    {:noreply, [], state}
-  end
-
   def recv(%Goaway{} = frame, state) do
     log_goaway(frame)
 
     {:stop, :normal, state}
-  end
-
-  def recv(%WindowUpdate{window_size_increment: inc}, state) do
-    flow = Connection.FlowControl.increment_window(state.flow_control, inc)
-    {:noreply, [], %{state | flow_control: flow}}
-  end
-
-  def recv(frame, state) do
-    """
-    Unknown RECV on connection
-    Frame: #{inspect(frame)}
-    State: #{inspect(state)}
-    """
-    |> Logger.info()
-
-    {:noreply, [], state}
-  end
-
-  def notify_settings_change(ref, old_settings, flow) do
-    %{initial_window_size: old_window} = old_settings
-    %{settings: settings} = flow
-
-    max_frame_size = settings.max_frame_size
-    new_window = settings.initial_window_size
-    window_diff = new_window - old_window
-
-    for stream_id <- flow.active_streams do
-      pid = Stream.via_tuple(ref, stream_id)
-      Stream.cast_recv(pid, {:settings_change, window_diff, max_frame_size})
-    end
   end
 
   defp do_send_headers(requests, state) when is_list(requests) do
@@ -280,7 +171,7 @@ defmodule Kadabra.Connection do
   end
 
   def handle_info({:closed, _pid}, %{config: config} = state) do
-    send(config.client, {:closed, config.supervisor})
+    Kernel.send(config.client, {:closed, config.supervisor})
     Task.start(fn -> Kadabra.Supervisor.stop(config.supervisor) end)
 
     {:stop, :normal, state}
@@ -299,119 +190,42 @@ defmodule Kadabra.Connection do
   end
 
   def handle_info({:push_promise, stream}, %{config: config} = state) do
-    send(config.client, {:push_promise, stream})
+    Kernel.send(config.client, {:push_promise, stream})
     {:noreply, [], state}
   end
 
   def handle_info({:recv, frame}, state) do
-    state = process(frame, state)
-    {:noreply, [], state}
+    case Processor.process(frame, %{config: config} = state) do
+      {:ok, state} ->
+        {:noreply, [], state}
+
+      {:connection_error, error, state} ->
+        code = Error.code(error)
+
+        bin =
+          state.flow_control.stream_id
+          |> Goaway.new(code)
+          |> Encodable.to_bin()
+
+        Socket.send(config.socket, bin)
+        {:stop, :normal, state}
+    end
   end
 
-  @spec process(frame, t) :: :ok
-  def process(bin, state) when is_binary(bin) do
-    Logger.info("Got binary: #{inspect(bin)}")
-    state
+  def send_protocol_error(state) do
+    send_goaway(state, "PROTOCOL_ERROR")
   end
 
-  def process(%Data{stream_id: 0}, state) do
-    # This is an error
-    state
+  def send_frame_size_error(state) do
+    send_goaway(state, "FRAME_SIZE_ERROR")
   end
 
-  def process(%Data{stream_id: stream_id} = frame, %{config: config} = state) do
-    send_window_update(config.socket, frame)
-
-    config.ref
-    |> Stream.via_tuple(stream_id)
-    |> Stream.cast_recv(frame)
-
-    state
-  end
-
-  def process(%Headers{stream_id: stream_id} = frame, %{config: config} = state) do
-    config.ref
-    |> Stream.via_tuple(stream_id)
-    |> Stream.call_recv(frame)
-
-    state
-  end
-
-  def process(%RstStream{} = frame, %{config: config} = state) do
-    pid = Stream.via_tuple(config.ref, frame.stream_id)
-    Stream.cast_recv(pid, frame)
-    state
-  end
-
-  def process(%Frame.Settings{} = frame, state) do
-    # Process immediately
-    {:noreply, [], state} = recv(frame, state)
-    state
-  end
-
-  def process(%PushPromise{stream_id: stream_id} = frame, state) do
-    %{config: config, flow_control: flow_control} = state
-    {:ok, pid} = StreamSupervisor.start_stream(config, flow_control, stream_id)
-
-    Stream.call_recv(pid, frame)
-
-    flow = Connection.FlowControl.add_active(flow_control, stream_id)
-
-    %{state | flow_control: flow}
-  end
-
-  def process(%Ping{} = frame, state) do
-    # Process immediately
-    recv(frame, state)
-    state
-  end
-
-  def process(%Goaway{} = frame, state) do
-    GenStage.cast(self(), {:recv, frame})
-    state
-  end
-
-  def process(%WindowUpdate{stream_id: 0} = frame, state) do
-    Stream.cast_recv(self(), frame)
-    state
-  end
-
-  def process(%WindowUpdate{stream_id: stream_id} = frame, state) do
-    pid = Stream.via_tuple(state.config.ref, stream_id)
-    Stream.cast_recv(pid, frame)
-    state
-  end
-
-  def process(%Continuation{stream_id: stream_id} = frame, state) do
-    pid = Stream.via_tuple(state.config.ref, stream_id)
-    Stream.call_recv(pid, frame)
-    state
-  end
-
-  def process(_error, state), do: state
-
-  def send_window_update(_socket, %Data{data: nil}), do: :ok
-
-  def send_window_update(_socket, %Data{data: ""}), do: :ok
-
-  def send_window_update(socket, %Data{stream_id: sid, data: data}) do
-    bin = data |> WindowUpdate.new() |> Encodable.to_bin()
-    Socket.send(socket, bin)
-
-    s_bin =
-      sid
-      |> WindowUpdate.new(byte_size(data))
-      |> Encodable.to_bin()
-
-    Socket.send(socket, s_bin)
-  end
-
-  def send_huge_window_update(socket) do
+  def send_goaway(%{config: config, flow_control: flow}, error) do
     bin =
-      0
-      |> Frame.WindowUpdate.new(2_000_000_000)
+      flow.stream_id
+      |> Frame.Goaway.new(Error.code(error))
       |> Encodable.to_bin()
 
-    Socket.send(socket, bin)
+    Socket.send(config.socket, bin)
   end
 end
